cqfd: have Dockerfile path in config file
Dockerfile path is hard coded as ".cqfd/docker/Dockerfile". Create a
field in build section in .cqfdrc to set a custom path for the
Dockerfile.
To set a new Dockerfile, add an entry 'dockerfile' in the build
section with the path to it.
A new docker image will be generated only when the 'cqfd init' command
is used.